new problem

+++ 10810.cpp
@@ -8,21 +8,21 @@
 using namespace std;
 
 #define N 500000
-//int v[2][N] = { };
-int v[N] = { };
-int u[N] = { };
+//long long v[2][N] = { };
+long long v[N] = { };
+long long u[N] = { };
 
-int mergeSortCount(int s, int e) {
+long long mergeSortCount(long long s, long long e) {
 
 	if (e == s + 1) { //no sorting
 		return 0;
 	}
 
-	int m = (s + e) / 2;
-	int re = mergeSortCount(s, m) + mergeSortCount(m, e);	// sort two halves
+	long long m = (s + e) / 2;
+	long long re = mergeSortCount(s, m) + mergeSortCount(m, e);	// sort two halves
 
 	// merge to buffer
-	int l = s, r = m, j = s;
+	long long l = s, r = m, j = s;
 	for (j = s; l < m && r < e; j++) {// while there are numbers on left and right
 		if (v[r] < v[l]) {
 			u[j] = v[r];
@@ -41,20 +41,20 @@ int mergeSortCount(int s, int e) {
 		u[j++] = v[l++];
 	}
 
-	for (int j=s;j<e;j++){
+	for (long long j=s;j<e;j++){
 		v[j]=u[j];
 	}
 	return re;
 }
 
 int main() {
-	int n;
-	while ((scanf("%d", &n) != EOF)&&n) {
-		for (int j = 0; j < n; j++) {
-			scanf("%d", v+j);
-			//printf("%d ", v[0][j]);
+	long long n;
+	while ((scanf("%lld", &n) != EOF)&&n) {
+		for (long long j = 0; j < n; j++) {
+			scanf("%lld", v+j);
+			//prlong longf("%d ", v[0][j]);
 		}
-		printf("%d\n", mergeSortCount(0, n));
+		printf("%lld\n", mergeSortCount(0, n));
 	}
 }
